refactor(examples): changes structure of examples
previously we had a wrapper in index.js around each example, which resulted in the example modules not using the actual data passed into a custom script.
with this change each example is a custom script on its own and bundled separately. The main bundle.js file is now used to link out to the actual examples

also changes the documentation to be in markdown format